+++ src/main/java/com/semi/main/product/ProductDAO.java
@@ -108,8 +108,4 @@ public class ProductDAO {
 		return sqlSession.selectOne(NAMESPACE+"dibsNum", productDTO);
 	}
 	
-	public Long imageCount(ProductDTO productDTO)throws Exception{
-		return sqlSession.selectOne(NAMESPACE+"imageCount", productDTO);
-	}
-	
 }
