+++ src/main/java/com/semi/main/product/ProductService.java
@@ -114,8 +114,5 @@ public class ProductService {
 	public Long dibsNum(ProductDTO productDTO)throws Exception{
 		return productDAO.dibsNum(productDTO);
 	}
-	
-	public Long imageCount(ProductDTO productDTO)throws Exception{
-		return productDAO.imageCount(productDTO);
-	}
+
 }
