+++ src/main/resources/database/config/database.properties
@@ -5,7 +5,7 @@
 #
 oracle.user=user02
 oracle.password=user02
-oracle.url=jdbc:oracle:thin:@13.125.180.64:1521:xe
+oracle.url=jdbc:oracle:thin:@52.79.171.94:1521:xe
 oracle.driver=oracle.jdbc.driver.OracleDriver
