+++ src/main/webapp/WEB-INF/views/product/detail.jsp
@@ -56,11 +56,70 @@
 	}
 	
 	.imgdiv{
-		width: 470px;
-		height: 450px;
+		width: 90%;
+		height: 95%;
 		
 	}
 	
+	.t{
+		font-weight : bold; 
+		font-size: 20px;
+	}
+	
+	.bi-heart-fill{
+		color: red;
+	}
+	
+	.bi{
+		margin: 4px;
+	}
+	
+	#wrap2{
+		width: 100%;
+		height: 1000px;
+
+	}
+		
+	.div3{
+		width: 65%;
+		height: 100%;
+		float: left;
+	}
+	.div4{
+		width: 35%;
+		height: 100%;
+		float: left;
+		
+	}
+	
+	.div5{
+		height:90%;
+		margin-top: 30px;
+		margin-right: 30px;
+	}
+	
+	.div6{
+		margin: 30px;
+		height: 90%;
+	}
+	
+	.div7{
+		height: 7%;
+		padding-top: 35px;
+	}
+	
+	.div8{
+		
+
+	}
+	#report{
+		margin-left: 300px;
+	}
+	
+	#rt{
+		color: #B7B5B5;
+		font-weight : bold;
+	}
 </style>
 </head>
 <body>
@@ -136,27 +195,40 @@
 		             			</span>
 							</div>
 							<div class="mt-2">
-								<span><i class="fa fa-heart-o me-2"></i>상품상태</span>
+								<span><i class="fa fa-heart-o me-2"></i>교환여부</span>
 		             			<span class="span ms-3">
 			             			<c:choose>
-								 	    <c:when test="${dto.proStatus eq 0}">
-								 			중고상품
+								 	    <c:when test="${dto.exchange eq 0}">
+								 			교환가능
 								 	    </c:when>
 								 	    <c:otherwise>
-								 			새상품
+								 			교환불가능
 								 	    </c:otherwise>
 							        </c:choose>
 		             			</span>
+		             			<span id="report"><a href="../member/reportadd?userNo=${dto.userNo}" id="rt">신고하기</a></span>
 							</div>
 							<div style="padding-top: 30px">
-		                        <button style="background-color: #15161D;" class="btn2" type="button"><div class="btn1">찜</div></button>
-		                        <button style="background-color: #fcca26;" class="btn2 ms-3" type="button"><div class="btn1">톡</div></button>
-		                        <button style="background-color: #D10024;" class="btn2 ms-3" type="button"><div class="btn3">바로구매</div></button>
+		                        <button style="background-color: #15161D;" class="btn2" id="dibs" type="button"><div class="t btn1"><i id="heart" class="bi bi-heart"></i>찜<span id="dibsNum" class="ms-2">${dibsNum}</span></div></button>
+		                        <button style="background-color: #fcca26;" class="btn2 ms-3" type="button"><div class="t btn1">톡</div></button>
+		                        <button id="buy" style="background-color: #D10024;" class="btn2 ms-3" type="button"><div class="t btn3">바로구매</div></button>
 	                         
 	                        </div>
 						</div>
 					</div>
 					
+				</div>
+				<div id="wrap2" class="border-top border-dark">
+					<div class="div3">
+						<div class="div5">
+							<div class="border-bottom div7"><h4 class="div8">상품정보</h4></div>
+						</div>
+					</div>
+					<div class="div4 border-start border-end">
+						<div class="div6">
+							<div class="border-bottom div7"><h4>판매자정보</h4></div>
+						</div>
+					</div>
 				</div>
 			</div>
 			<!-- /container -->
