+++ src/main/webapp/resources/js/detailLogout.js
@@ -1,6 +1,6 @@
 	let proNo=$('#add').attr("data-add-pnum");
 	
-	$('#dibs').click(function(){
-		alert("로그인 상태에서만 찜하기가 가능합니다.");
+	$('.btn2').click(function(){
+		alert("로그인 상태에서만 가능합니다.");
 		location.href="../member/login";
 	})
